% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winn.R
\name{scale_by_batch}
\alias{scale_by_batch}
\title{Scale Data by batch}
\usage{
scale_by_batch(data, batch)
}
\arguments{
\item{data}{A numeric matrix with rows representing metabolites and columns representing samples.}

\item{batch}{A numeric vector indicating the batch (or segment) assignment for each sample.}
}
\value{
A numeric matrix of scaled intensities.
}
\description{
This function scales the values for each metabolite within each batch by subtracting the batch mean and dividing by the batch standard deviation.
}
\examples{
your_data_matrix <- matrix(rnorm(200, mean = 100, sd = 15), nrow = 20)
batch <- rep(1:4, length.out = ncol(your_data_matrix))
scaled_data <- scale_by_batch(your_data_matrix, batch)
}
