% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winn.R
\name{adjust_outliers_mad}
\alias{adjust_outliers_mad}
\title{Adjust Outliers in Data Matrix Using MAD}
\usage{
adjust_outliers_mad(data)
}
\arguments{
\item{data}{A numeric matrix or data frame where rows represent metabolites and columns represent samples.}
}
\value{
A numeric matrix with adjusted outliers.
}
\description{
This function adjusts outliers in a data matrix on a per-metabolite basis using the Median Absolute Deviation (MAD).
}
\examples{
your_data_matrix <- matrix(rnorm(200, mean = 100, sd = 15), nrow = 20)
adjusted_data <- adjust_outliers_mad(your_data_matrix)
}
